+++ index.ts
@@ -1,0 +1,21 @@
+import PayFast from './src/Payfast'
+import PayFastSaveCard from './src/SaveCard'
+import PayFastUpdateCard from './src/UpdateCard'
+import usePayFast from './src/hooks'
+import {
+  PayFastMerchantDetails,
+  PayFastSubscriptionDetails,
+  PayFastSubscriptionFrequency,
+  PayFastTransactionDetails
+} from './src/types'
+
+export {
+  PayFastMerchantDetails,
+  PayFastSaveCard,
+  PayFastSubscriptionDetails,
+  PayFastSubscriptionFrequency,
+  PayFastTransactionDetails,
+  PayFastUpdateCard,
+  usePayFast
+}
+export default PayFast