+++ index.tsx
@@ -1,4 +1,0 @@
-import PayFast, { PayFastTransactionDetails } from './src/Payfast'
-
-export { PayFastTransactionDetails }
-export default PayFast